# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Sözlük API
  version: 2.0.0-m2
  description: Milestone 2 REST contract. All dates are ISO 8601 UTC strings.
servers:
  - url: /
security:
  - sessionCookie: []
    csrfHeader: []
...
paths:
...
  /api/v1/admin/agent-runtime/events:
    get:
      operationId: streamAdminAgentRuntimeEvents
      tags: [Admin]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: afterId
          in: query
          schema: { type: string, pattern: "^[0-9]{1,20}$" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: poll
          in: query
          schema: { type: integer, enum: [1] }
      responses:
        {
          "200": { $ref: "#/components/responses/RuntimeEventStream" },
          "403": { $ref: "#/components/responses/Error" },
          "422": { $ref: "#/components/responses/ValidationError" },
          default: { $ref: "#/components/responses/Error" },
        }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ajan_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the session cookie and a validated same-origin Origin header for every POST, PUT, PATCH and DELETE request.
    runtimeBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-agent-runtime-token
      description: Hash-stored agent runtime credential. It is never available to the Codex child process and cannot be combined with a browser session.
...
  headers:
    RequestId:
      {
        required: true,
        schema: { type: string, format: uuid },
        description: Preserves a valid incoming X-Request-Id or contains a generated UUID.,
      }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [data, requestId]
      properties: { data: {}, requestId: { type: string, format: uuid } }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            fieldErrors:
              { type: object, additionalProperties: { type: array, items: { type: string } } }
            requestId: { type: string, format: uuid }
...
  responses:
...
    RuntimeEventStream:
      description: SSE stream by default; JSON safe-event array when poll=1 is supplied
      headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
      content:
        text/event-stream: { schema: { type: string } }
        application/json: { schema: { $ref: "#/components/schemas/SuccessEnvelope" } }
    ValidationError:
      description: Validation failed
      headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
...
    Error:
      description: Authentication, authorization, not-found, conflict or unexpected error. Production responses never expose a stack trace.
      headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }